# Make the ContentItemPresenter much faster