--- a/app/presenters/queries/content_item_presenter.rb
+++ b/app/presenters/queries/content_item_presenter.rb
@@ -1,182 +1,149 @@
+# This presenter has been carefully written to run quickly. Please be careful
+# if editing its behaviour and make sure to compare benchmarks.
 module Presenters
   module Queries
     class ContentItemPresenter
-      attr_accessor :order, :offset, :limit
+      attr_accessor :scope, :fields, :order, :limit, :offset
+
+      DEFAULT_FIELDS = [
+        *ContentItem::TOP_LEVEL_FIELDS,
+        :publication_state,
+        :user_facing_version,
+        :base_path,
+        :locale,
+        :lock_version,
+        :internal_name,
+      ]
+
+      def self.present_many(scope, params = {})
+        new(scope, params).present_many
+      end
 
       def self.present(content_item)
         translation = Translation.find_by!(content_item: content_item)
 
-        content_items = ContentItem.where(content_id: content_item.content_id)
-        content_items = Translation.filter(content_items, locale: translation.locale)
+        scope = ContentItem.where(content_id: content_item.content_id)
+        scope = Translation.filter(scope, locale: translation.locale)
 
-        present_many(content_items).first
+        present_many(scope).first
       end
 
-      def self.present_many(content_item_scope, fields: nil, order: { public_updated_at: :desc }, offset: 0, limit: nil, locale: nil)
-        presenter = new(content_item_scope, fields: fields, order: order, limit: limit, offset: offset, locale: locale)
-        presenter.present
+      def initialize(scope, params = {})
+        self.scope = scope
+        self.fields = (params[:fields] || DEFAULT_FIELDS).map(&:to_sym)
+        self.order = params[:order]
+        self.limit = params[:limit]
+        self.offset = params[:offset]
       end
 
-      def initialize(content_item_scope, fields: nil, order: {}, limit: nil, offset: 0, locale: nil)
-        self.content_item_scope = content_item_scope
-        self.fields = fields
-        self.order = order
-        self.limit = limit
-        self.offset = offset
-        self.locale = locale
-      end
+      def present_many
+        scope = ::Queries::GetLatest.call(self.scope)
+        scope = join_supporting_objects(scope)
+        scope = order_and_paginate(scope)
 
-      def present
-        group_items(groups).compact
+        extract_fields(scope)
       end
 
       def total
-        ActiveRecord::Base.connection.execute(total_query).first["count"].to_i
+        ::Queries::GetLatest.call(self.scope).count
       end
 
     private
 
-      attr_accessor :content_item_scope, :fields, :locale
-
-      def total_query
-        sql = content_item_scope.select(:content_id)
-        sql = sql.group(:content_id) unless locale == "all"
-        "SELECT COUNT(*) FROM (#{sql.to_sql}) total"
-      end
-
-      def groups
-        ActiveRecord::Base.connection.execute(aggregated_sql(scope.to_sql))
-      end
-
-      def scope
-        scope = join_supporting_objects(content_item_scope)
-        select_fields(scope)
-      end
-
-      def aggregated_sql(sql)
-        <<-END.strip_heredoc
-          #{aggregated_query(sql)}
-          #{aggregated_order}
-          OFFSET #{offset}
-          #{aggregated_limit}
-        END
-      end
-
-      def aggregated_query(sql)
-        <<-END.strip_heredoc
-          SELECT json_agg(json_rows) FROM (
-            SELECT row_to_json(item) json_item FROM (#{sql}) item
-          ) json_rows
-          GROUP BY json_item->>'content_id', json_item->>'locale'
-        END
-      end
-
-      def aggregated_order
-        "ORDER BY MAX(json_item->>'#{order_field}') #{direction}"
-      end
-
-      def order_field
-        order.keys.first || :public_updated_at
-      end
-
-      def direction
-        order.values.first || :desc
-      end
-
-      def aggregated_limit
-        "LIMIT #{limit}" if limit
-      end
-
-      def group_items(groups)
-        groups.map do |raw_group|
-          items = JSON.parse(raw_group["json_agg"]).map { |g| g["json_item"] }
-
-          draft = detect_draft(items)
-          live = detect_live(items)
-
-          most_recent_item = draft || live
-          next unless most_recent_item
-
-          most_recent_item["description"] = most_recent_item["description"]["value"] if most_recent_item["description"]
-
-          if output_fields.include?("internal_name")
-            details = most_recent_item.fetch("details")
-            most_recent_item["internal_name"] = details["internal_name"] || most_recent_item.fetch("title")
-          end
-
-          most_recent_item["publication_state"] = publication_state(draft, live)
-          most_recent_item["lock_version"] = most_recent_item.fetch("lock_version").to_i
-
-          if live
-            most_recent_item["live_version"] = live.fetch("lock_version").to_i
-          end
-
-          most_recent_item.slice(*output_fields)
-        end
-      end
-
       def join_supporting_objects(scope)
         scope = State.join_content_items(scope)
+        scope = UserFacingVersion.join_content_items(scope)
         scope = Translation.join_content_items(scope)
         scope = Location.join_content_items(scope)
-        scope = UserFacingVersion.join_content_items(scope)
-        scope = LockVersion.join_content_items(scope)
 
+        LockVersion.join_content_items(scope)
+      end
+
+      def order_and_paginate(scope)
+        scope = scope.order(order.to_a.join(" ")) if order
+        scope = scope.limit(limit) if limit
+        scope = scope.offset(offset) if offset
         scope
       end
 
-      def select_fields(scope)
-        ordering_fields = order.keys.map(&:to_s)
-        scope.select(
-          *ContentItem::TOP_LEVEL_FIELDS,
-          "content_id",
-          "states.name as state_name",
-          "lock_versions.number as lock_version",
-          "translations.locale",
-          "locations.base_path",
-          *ordering_fields,
-        )
-      end
-
-      def output_fields
-        if fields
-          output_fields = fields
-        else
-          additional_fields = %w(
-            locale
-            base_path
-            lock_version
-            publication_state
-            live_version
-          )
-
-          output_fields = ContentItem::TOP_LEVEL_FIELDS + additional_fields
+      def extract_fields(scope)
+        fields_to_select = fields.map do |field|
+          case field
+          when :publication_state
+            "#{publication_state_sql} AS publication_state"
+          when :user_facing_version
+            "user_facing_versions.number AS user_facing_version"
+          when :lock_version
+            "lock_versions.number AS lock_version"
+          when :description
+            "description->>'value' AS description"
+          when :internal_name
+            "#{internal_name_sql} AS internal_name"
+          when :public_updated_at
+            "to_char(public_updated_at, '#{iso8601_sql}') as public_updated_at"
+          else
+            field
+          end
         end
 
-        output_fields.map(&:to_s)
+        scope = scope.select(*fields_to_select)
+        results = evaluate_query(scope)
+
+        parsing_enumerator(results)
       end
 
-      def publication_state(draft, live)
-        draft_lock_version = draft.fetch("lock_version") if draft
-        live_lock_version = live.fetch("lock_version") if live
+      def publication_state_sql
+        <<-SQL
+          CASE WHEN (user_facing_versions.number > 1 AND states.name = 'draft') THEN
+            'redrafted'
+          WHEN (states.name = 'published') THEN
+            'live'
+          ELSE
+            states.name
+          END
+        SQL
+      end
 
-        if draft_lock_version && live_lock_version && (draft_lock_version > live_lock_version)
-          "redrafted"
-        elsif live
-          "live"
-        elsif draft
-          "draft"
-        else
-          raise "Something unexpected happened"
+      def iso8601_sql
+        "YYYY-MM-DD\"T\"HH24:MI:SS\"Z\""
+      end
+
+      # This returns the internal_name from the details hash if it is present,
+      # otherwise it falls back to the content item's title.
+      def internal_name_sql
+        "COALESCE(details->>'internal_name', title) "
+      end
+
+      def parsing_enumerator(results)
+        json_columns = %w(details routes redirects need_ids)
+        int_columns = %w(user_facing_version lock_version)
+
+        Enumerator.new do |yielder|
+          results.each do |result|
+            json_columns.each { |c| parse_json_column(result, c) }
+            int_columns.each { |c| parse_int_column(result, c) }
+
+            yielder.yield result
+          end
         end
       end
 
-      def detect_draft(items)
-        items.detect { |i| i.fetch("state_name") == "draft" }
+      def parse_json_column(result, column)
+        return unless result.key?(column)
+        result[column] = JSON.parse(result[column])
       end
 
-      def detect_live(items)
-        items.detect { |i| i.fetch("state_name") == "published" }
+      def parse_int_column(result, column)
+        return unless result.key?(column)
+        result[column] = result[column].to_i
+      end
+
+      # It is substantially faster to evaluate directly against Postgres than
+      # to use: #as_json, #pluck or even ActiveRecord::Base.connection.execute
+      # The reason for this is that the results of this query don't have to go
+      # through as many abstraction layers. This is closer to the metal.
+      def evaluate_query(scope)
+        ActiveRecord::Base.connection.raw_connection.exec(scope.to_sql)
       end
     end
   end