… spec/queries/get_content_spec.rb
… spec/queries/get_latest_spec.rb
… spec/service_consumers/pact_helper.rb