https://trello.com/c/1wLclYR2/668-improve-efficiency-of-content-item-presenter

**Note:** Tests won't pass until this is merged: https://github.com/alphagov/gds-api-adapters/pull/488

Previously, the ContentItemPresenter could take a long time,
even with pagination and this was causing timeouts across a
number of projects and general instability. It was also
slowing down put content requests because the response goes
through the presenter.

This work effectively rewrites this presenter to work much
more quickly. It relies on `Queries::GetLatest` which is an
example of a "get rows with max value per group" query.
There are further improvements that could be made, perhaps
with how the count is computed, but these changes are
probably good enough for now.

The old presenter made it so that public_updated_at was
always output, regardless of whether this field was
requested or not. I have fixed this. It should now be
explicitly requested, just like any other field. In
addition to this, I made it so that the presenter defaults
to return all fields. The presenter is now quick enough to
do this and I think it serves as good documentation as to
which fields are available.

Finally, here are the results of some quick benchmarking
in development. Times are in milliseconds:

```
/v2/linkables?document_type=organisation
new: 95, 100, 96, 97, 95 (avg 97)
old: 553, 302, 535, 519, 563 (avg 494)
average: 509% as fast

/v2/linkables?document_type=policy
new: 39, 44, 37, 44, 38 (avg 40)
old: 109, 91, 92, 197, 210 (avg 140)
average: 350% as fast

/v2/content?document_type=organisation&fields[]=title&fields[]=base_path
new: 84, 87, 83, 90, 87 (avg 86)
old: 320, 361, 374, 351, 368 (avg 355)
average: 416% as fast

/v2/content?document_type=cma_case&fields[]=description&per_page=1000
new: 116, 111, 109, 114, 117 (avg 113)
old: 985, 678, 964, 953, 730 (avg 862)
average: 763% as fast

/v2/content?document_type=travel_advice&fields[]=details&per_page=3
new: 98, 99, 98, 105, 107 (avg 101)
old: 1363, 1048, 1020, 1025, 1030 (avg 1097)
average: 1086% as fast

PUT /v2/content/08d48cdd-6b50-43ff-a53b-beab47f4aab0 (travel advice index)
new: 904, 780, 804, 660, 883 (avg 806)
old: 3601, 2483, 2683, 2542, 2592 (avg 2780)
average: 345% as fast
```

Note: This last case was particularly problematic because the default
gds-api-adapters timeout is 4 seconds, which the old implementation
was close to hitting.

## Benchmarks graph

![benchmark](https://cloud.githubusercontent.com/assets/892251/14531374/fc9ee7c0-0254-11e6-817f-aec7c95c7f86.png)